Rename expr to expression (#473)
* Rename RecordExpr to RecordExpression
* Rename ExprParser to ExpressionParser
