+++ core/expressionParser/helpers/evaluator.js
