+++ webapp/commonComponents/expression/nodes/editButtons.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Expression from '@core/exprParser/expression'
+import Expression from '@core/expressionParser/expression'
 
 import { useI18n } from '../../hooks'
