+++ webapp/commonComponents/expression/nodes/expressionNode.js
@@ -9,7 +9,7 @@ import Literal from './literal'
 import Logical from './logical'
 import Member from './member'
 
-import Expression from '@core/exprParser/expression'
+import Expression from '@core/expressionParser/expression'
 
 const components = {
   [Expression.types.Identifier]: Identifier,
