+++ webapp/commonComponents/expression/nodes/member.js
@@ -1,6 +1,6 @@
 import React from 'react'
 
-import Expression from '@core/exprParser/expression'
+import Expression from '@core/expressionParser/expression'
 
 import Identifier from './identifier'
